move fetch_attribute_value
